test(web): fold HTTP status passthrough tests into existing settings test file
